# ShopApp/src/main/java/ru/got/shop/dto/ResponseWrapperAdsCommentDto.java
package ru.got.shop.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;

import javax.validation.Valid;
import java.util.List;

@AllArgsConstructor
@Builder
@Data
@Valid
public class ResponseWrapperAdsCommentDto {

    @JsonProperty("count")
    private Integer count;

    @JsonProperty("results")
    @Valid
    private List<AdCommentDto> results;
}